+++ config/initializers/home_control.rb
@@ -1,7 +1,7 @@
 require "ostruct"
 Rails.application.configure do
   config.home_control = OpenStruct.new(
-    version: "3.5.0",
+    version: "3.5.1",
     authentication: {
       name: ENV["ADMIN_USERNAME"],
       password: ENV["ADMIN_PASSWORD"]